re-jig a lot of stuff for organization
