+++ packages/thirdweb/src/extensions/marketplace/direct-listings/direct-listings.test.ts
@@ -1,0 +1,225 @@
+import { beforeAll, describe, expect, it } from "vitest";
+import { ANVIL_CHAIN } from "../../../../test/src/chains.js";
+import { TEST_CLIENT } from "../../../../test/src/test-clients.js";
+import {
+  TEST_ACCOUNT_A,
+  TEST_ACCOUNT_B,
+} from "../../../../test/src/test-wallets.js";
+import {
+  type ThirdwebContract,
+  getContract,
+} from "../../../contract/contract.js";
+import { parseEventLogs } from "../../../event/actions/parse-logs.js";
+import { balanceOf } from "../../../exports/extensions/erc721.js";
+import { sendAndConfirmTransaction } from "../../../transaction/actions/send-and-confirm-transaction.js";
+import { approve } from "../../erc721/__generated__/IERC721A/write/approve.js";
+import { tokensMintedEvent } from "../../erc721/__generated__/IMintableERC721/events/TokensMinted.js";
+import { mintTo } from "../../erc721/write/mintTo.js";
+import { deployERC721Contract } from "../../prebuilts/deploy-erc721.js";
+import { deployMarketplaceContract } from "../../prebuilts/deploy-marketplace.js";
+import { newListingEvent } from "../__generated__/IDirectListings/events/NewListing.js";
+import { totalListings } from "../__generated__/IDirectListings/read/totalListings.js";
+import { getAllListings } from "./read/getAllListings.js";
+import { getAllValidListings } from "./read/getAllValidListings.js";
+import { getListing } from "./read/getListing.js";
+import { isListingValid } from "./utils.js";
+import { buyFromListing } from "./write/buyFromListing.js";
+import { createListing } from "./write/createListing.js";
+
+describe("Marketplace: Direct Listings", () => {
+  let nftTokenId: bigint;
+  let marketplaceContract: ThirdwebContract;
+  let erc721Contract: ThirdwebContract;
+  beforeAll(async () => {
+    const marketplaceAddress = await deployMarketplaceContract({
+      account: TEST_ACCOUNT_A,
+      chain: ANVIL_CHAIN,
+      client: TEST_CLIENT,
+      params: {
+        name: "TestMarketPlace",
+      },
+    });
+    marketplaceContract = getContract({
+      address: marketplaceAddress,
+      client: TEST_CLIENT,
+      chain: ANVIL_CHAIN,
+    });
+
+    // also deploy an ERC721 contract
+    const erc721Address = await deployERC721Contract({
+      type: "TokenERC721",
+      account: TEST_ACCOUNT_A,
+      chain: ANVIL_CHAIN,
+      client: TEST_CLIENT,
+      params: {
+        name: "TestERC721",
+      },
+    });
+
+    erc721Contract = getContract({
+      address: erc721Address,
+      client: TEST_CLIENT,
+      chain: ANVIL_CHAIN,
+    });
+
+    const mintTransaction = mintTo({
+      contract: erc721Contract,
+      to: TEST_ACCOUNT_A.address,
+      nft: { name: "Test:ERC721:DirectListing" },
+    });
+    const receipt = await sendAndConfirmTransaction({
+      transaction: mintTransaction,
+      account: TEST_ACCOUNT_A,
+    });
+
+    const mintEvents = parseEventLogs({
+      events: [tokensMintedEvent()],
+      logs: receipt.logs,
+    });
+
+    expect(mintEvents.length).toBe(1);
+    expect(mintEvents[0]?.args.tokenIdMinted).toBeDefined();
+
+    nftTokenId = mintEvents[0]?.args.tokenIdMinted as bigint;
+    // does a lot of stuff, this may take a while
+  }, 120_000);
+
+  it("should work for basic listings (Native Currency)", async () => {
+    // listings should be 0 length to start
+    const listings = await getAllListings({
+      contract: marketplaceContract,
+    });
+    expect(listings.length).toBe(0);
+    // oh and so should totalListings
+    expect(await totalListings({ contract: marketplaceContract })).toBe(0n);
+
+    // approve first
+    const approveTx = approve({
+      contract: erc721Contract,
+      to: marketplaceContract.address,
+      tokenId: nftTokenId,
+    });
+
+    await sendAndConfirmTransaction({
+      transaction: approveTx,
+      account: TEST_ACCOUNT_A,
+    });
+
+    const transaction = createListing({
+      contract: marketplaceContract,
+      assetContractAddress: erc721Contract.address,
+      tokenId: nftTokenId,
+      pricePerToken: "1",
+    });
+    const receipt = await sendAndConfirmTransaction({
+      transaction,
+      account: TEST_ACCOUNT_A,
+    });
+
+    const listingEvents = parseEventLogs({
+      events: [newListingEvent()],
+      logs: receipt.logs,
+    });
+
+    expect(listingEvents.length).toBe(1);
+
+    // biome-ignore lint/style/noNonNullAssertion: OK in tests
+    const listingEvent = listingEvents[0]!;
+
+    expect(listingEvent.args.listingCreator).toBe(TEST_ACCOUNT_A.address);
+    expect(listingEvent.args.assetContract).toBe(erc721Contract.address);
+
+    // at this point listings should be 1
+    const listingsAfter = await getAllListings({
+      contract: marketplaceContract,
+    });
+    expect(listingsAfter.length).toBe(1);
+    // valid listings should also be 1!
+    const validListings = await getAllValidListings({
+      contract: marketplaceContract,
+    });
+    expect(validListings.length).toBe(1);
+    // and totalListings should be 1
+    expect(await totalListings({ contract: marketplaceContract })).toBe(1n);
+
+    // explicitly retrieve the listing!
+    const listing = await getListing({
+      contract: marketplaceContract,
+      listingId: listingEvent.args.listingId,
+    });
+
+    expect(listing).toBeDefined();
+    expect(listing.status).toBe("ACTIVE");
+    expect(listing.creatorAddress).toBe(TEST_ACCOUNT_A.address);
+    expect(listing.assetContractAddress).toBe(erc721Contract.address);
+    expect(listing.tokenId).toBe(nftTokenId);
+    expect(listing.currencyValuePerToken).toMatchInlineSnapshot(`
+    {
+      "decimals": 18,
+      "displayValue": "1",
+      "name": "Anvil Ether",
+      "symbol": "ETH",
+      "value": 1000000000000000000n,
+    }
+  `);
+    expect(listing.asset).toMatchInlineSnapshot(`
+    {
+      "id": 0n,
+      "metadata": {
+        "name": "Test:ERC721:DirectListing",
+      },
+      "owner": null,
+      "tokenURI": "ipfs://QmewoGwuooC1Vno1ZTnmKmEUaimDygraR1fEEHdTeraF34/0",
+      "type": "ERC721",
+    }
+  `);
+
+    // check the listing is valid
+    const listingValidity = await isListingValid({
+      listing,
+      contract: marketplaceContract,
+      quantity: 1n,
+    });
+
+    expect(listingValidity).toMatchInlineSnapshot(`
+    {
+      "valid": true,
+    }
+  `);
+
+    // expect the buyer to have an initial balance of 0
+    await expect(
+      balanceOf({
+        contract: erc721Contract,
+        owner: TEST_ACCOUNT_B.address,
+      }),
+    ).resolves.toBe(0n);
+
+    const buyTx = buyFromListing({
+      contract: marketplaceContract,
+      listingId: listingEvent.args.listingId,
+      recipient: TEST_ACCOUNT_B.address,
+      quantity: 1n,
+    });
+
+    await sendAndConfirmTransaction({
+      transaction: buyTx,
+      account: TEST_ACCOUNT_B,
+    });
+
+    // expect the buyer to have a new balance of 1
+    await expect(
+      balanceOf({
+        contract: erc721Contract,
+        owner: TEST_ACCOUNT_B.address,
+      }),
+    ).resolves.toBe(1n);
+    // expect the seller to no longer have the token
+    await expect(
+      balanceOf({
+        contract: erc721Contract,
+        owner: TEST_ACCOUNT_A.address,
+      }),
+    ).resolves.toBe(0n);
+  });
+});
