+++ packages/thirdweb/src/extensions/marketplace/english-auctions/read/getAllAuctions.ts
@@ -2,14 +2,15 @@ import { eth_getBlockByNumber } from "../../../../rpc/actions/eth_getBlockByNumb
 import { getRpcClient } from "../../../../rpc/rpc.js";
 import type { BaseTransactionOptions } from "../../../../transaction/types.js";
 import { min } from "../../../../utils/bigint.js";
-import { getAllValidListings as getAllValidListingGenerated } from "../../__generated__/IDirectListings/read/getAllValidListings.js";
-import { totalListings } from "../../__generated__/IDirectListings/read/totalListings.js";
-import type { DirectListing } from "../../types.js";
-import { getAllInBatches, mapDirectListing } from "../../utils.js";
+import { getAllAuctions as getAllAuctionsGenerated } from "../../__generated__/IEnglishAuctions/read/getAllAuctions.js";
+import { totalAuctions } from "../../__generated__/IEnglishAuctions/read/totalAuctions.js";
+import { getAllInBatches } from "../../utils.js";
+import type { EnglishAuction } from "../types.js";
+import { mapEnglishAuction } from "../utils.js";
 
 const DEFAULT_QUERY_ALL_COUNT = 100n;
 
-export type GetAllValidListingParams = {
+export type GetAllAuctionParams = {
   /**
    * The start index of the listings to retrieve.
    * @default 0
@@ -23,22 +24,22 @@ export type GetAllValidListingParams = {
 };
 
 /**
- * Retrieves all valid direct listings based on the provided options.
- * @param options - The options for retrieving the valid listing.
- * @returns A promise that resolves to the direct listings array.
+ * Retrieves all auctions based on the provided options.
+ * @param options - The options for retrieving the auctions.
+ * @returns A promise that resolves to the auctions array.
  * @extension MARKETPLACE
  * @example
  *
  * ```ts
- * import { getAllValidListings } from "thirdweb/extensions/marketplace";
+ * import { getAllAuctions } from "thirdweb/extensions/marketplace";
  *
- * const validListings = await getAllValidListings({ contract, start: 0, count: 10 });
+ * const listings = await getAllAuctions({ contract, start: 0, count: 10 });
  * ```
  */
-export async function getAllValidListings(
-  options: BaseTransactionOptions<GetAllValidListingParams>,
-): Promise<DirectListing[]> {
-  const totalCount = await totalListings(options);
+export async function getAllAuctions(
+  options: BaseTransactionOptions<GetAllAuctionParams>,
+): Promise<EnglishAuction[]> {
+  const totalCount = await totalAuctions(options);
   // if the totalListingCount is 0, return an empty array and skip all other work
   if (totalCount === 0n) {
     return [];
@@ -49,14 +50,10 @@ export async function getAllValidListings(
   const end = min(totalCount, start + count);
 
   const rpcClient = getRpcClient(options.contract);
-  const [rawListings, latestBlock] = await Promise.all([
+  const [rawAuctions, latestBlock] = await Promise.all([
     getAllInBatches(
       (startId, endId) =>
-        getAllValidListingGenerated({
-          contract: options.contract,
-          startId,
-          endId,
-        }),
+        getAllAuctionsGenerated({ contract: options.contract, startId, endId }),
       {
         start,
         end,
@@ -71,11 +68,11 @@ export async function getAllValidListings(
   ]);
 
   return await Promise.all(
-    rawListings.map((rawListing) =>
-      mapDirectListing({
+    rawAuctions.map((rawAuction) =>
+      mapEnglishAuction({
         contract: options.contract,
         latestBlock,
-        rawListing,
+        rawAuction,
       }),
     ),
   );
