+++ packages/thirdweb/src/extensions/marketplace/offers/read/getAllOffers.ts
@@ -2,43 +2,44 @@ import { eth_getBlockByNumber } from "../../../../rpc/actions/eth_getBlockByNumb
 import { getRpcClient } from "../../../../rpc/rpc.js";
 import type { BaseTransactionOptions } from "../../../../transaction/types.js";
 import { min } from "../../../../utils/bigint.js";
-import { getAllListings as getAllListingGenerated } from "../../__generated__/IDirectListings/read/getAllListings.js";
-import { totalListings } from "../../__generated__/IDirectListings/read/totalListings.js";
-import type { DirectListing } from "../../types.js";
-import { getAllInBatches, mapDirectListing } from "../../utils.js";
+import { getAllOffers as getAllOffersGenerated } from "../../__generated__/IOffers/read/getAllOffers.js";
+import { totalOffers } from "../../__generated__/IOffers/read/totalOffers.js";
+import { getAllInBatches } from "../../utils.js";
+import type { Offer } from "../types.js";
+import { mapOffer } from "../utils.js";
 
 const DEFAULT_QUERY_ALL_COUNT = 100n;
 
-export type GetAllListingParams = {
+export type GetAllOffersParams = {
   /**
-   * The start index of the listings to retrieve.
+   * The start index of the offers to retrieve.
    * @default 0
    */
   start?: number;
   /**
-   * The number of listings to retrieve.
+   * The number of offers to retrieve.
    * @default 100
    */
   count?: bigint;
 };
 
 /**
- * Retrieves all direct listings based on the provided options.
- * @param options - The options for retrieving the listing.
- * @returns A promise that resolves to the direct listings array.
+ * Retrieves all offers based on the provided options.
+ * @param options - The options for retrieving the offers.
+ * @returns A promise that resolves to the offers array.
  * @extension MARKETPLACE
  * @example
  *
  * ```ts
- * import { getAllListings } from "thirdweb/extensions/marketplace";
+ * import { getAllOffers } from "thirdweb/extensions/marketplace";
  *
- * const listings = await getAllListings({ contract, start: 0, count: 10 });
+ * const listings = await getAllOffers({ contract, start: 0, count: 10 });
  * ```
  */
-export async function getAllListings(
-  options: BaseTransactionOptions<GetAllListingParams>,
-): Promise<DirectListing[]> {
-  const totalCount = await totalListings(options);
+export async function getAllOffers(
+  options: BaseTransactionOptions<GetAllOffersParams>,
+): Promise<Offer[]> {
+  const totalCount = await totalOffers(options);
   // if the totalListingCount is 0, return an empty array and skip all other work
   if (totalCount === 0n) {
     return [];
@@ -49,10 +50,10 @@ export async function getAllListings(
   const end = min(totalCount, start + count);
 
   const rpcClient = getRpcClient(options.contract);
-  const [rawListings, latestBlock] = await Promise.all([
+  const [rawOffers, latestBlock] = await Promise.all([
     getAllInBatches(
       (startId, endId) =>
-        getAllListingGenerated({ contract: options.contract, startId, endId }),
+        getAllOffersGenerated({ contract: options.contract, startId, endId }),
       {
         start,
         end,
@@ -67,11 +68,11 @@ export async function getAllListings(
   ]);
 
   return await Promise.all(
-    rawListings.map((rawListing) =>
-      mapDirectListing({
+    rawOffers.map((rawOffer) =>
+      mapOffer({
         contract: options.contract,
         latestBlock,
-        rawListing,
+        rawOffer,
       }),
     ),
   );
